backend/requirements.txt
# Database
psycopg2-binary>=2.9.9

# Message Broker
pika>=1.3.2

# API Server
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.9

# Media Processing
ffmpeg-python>=0.2.0
opencv-python>=4.10.0
Pillow>=10.4.0

# Audio Analysis
librosa>=0.10.2
soundfile>=0.12.1

# ML / AI
torch>=2.3.0
torchvision>=0.18.0
transformers>=4.44.0
accelerate>=0.33.0
qwen-vl-utils>=0.0.2
sentence-transformers>=2.7.0
scikit-learn>=1.3.0

# Utilities
numpy>=1.26.0
python-dotenv>=1.0.0
